+++ zarrs/src/array/codec/array_to_bytes/vlen_array/vlen_array_codec.rs
@@ -1,0 +1,3 @@
+use crate::array::codec::array_to_bytes::vlen_v2::vlen_v2_macros;
+
+vlen_v2_macros::vlen_v2_codec!(VlenArrayCodec, "vlen-array");